Create MINSUM.vba
THIS IS A VBA CUSTOM FUNCTION FOR MICROSOFT EXCEL:
MINSUM(RANGE)
Calculates the minimum possible aggregate value by comparing the sum of the values in the selected cells. Example: the user selected range "A1:F1". The values in the selected cells are:
A1 = 10
B1 = -3
C1 = 7
D1 = -20
E1 = -5
F1 = 35

MINSUM(A1:F1) returns the minimum possible aggregate value obtained by comparing the value in A1 vs. sum(A1:B1) vs. sum(A1:C1) vs. sum(A1:D1) vs. sum(A1:E1) vs. sum(A1:F1). In this specific instance, the result is -11.

This function can be used for calculating the minimum level of a projected cumulative cash flow within a specified time frame. I used it for telling a financial model when to project a capital distribution in the future. This can be done by comparing the result of this function to the minimum bank balance value allowed (i.e 250K USD). When the result of the function is greater than 250k USD, a cash distribution can be expected to be made.

You could also make this a public function if you need to use it in other VBA modules.